added bigquery_table - moved bigquery entities to its own file added Table class and is_metadata to table field
